Very Early in Dev Reform Decisions
(DOES NOT WORK)

Main idea is:
When decision pressed it will add +5 to a variable while removing -5 from another. For this to be functional, the variable who is being reduced has to have a minimum of 5 or above. This however, is not consistent in the game.

Some decisions just work despite a variable not having more than 5 and none of the decisions don't take -5 only add +5.