fixed the bugs
this bug was due to the ideas file. no idea why, all i needed to was delte the original and create a clone and then it worked. maybe the ecoding was changed or something, no idea. but ideas are functioning well. that took way to long to figure out.

if in the future, we get the add... error again, create the same idea in another file and if it works then. its the file itself thats bugged out