Getting There
Just a disclaimer, Knightmare designer doesn't work at all but that is ok because I am working on it right now. On a more important note, I fixed the problem with tanks and 2 more problems have arisen that I am struggling to fix. On that note, I am publishing it semi-broken still because my computer needs a defrag and I don't want to re find the solution again in case it messes with the mod files.
Also may have accidentally deleted some png files, we will see once this goes through.