Refinement of Tanks
Just renamed some stuff to base game so I could use the GFX and localization. The tank designer is like 99% done just one little bug or thing I want to try and fix because it makes the designer not feel right. Tons of localization missing and even more GFX missing but the code is basically all there.
Oh, still doesn't have a tech tree as well with everything unlocking at the start for testing purposes.